A simple change to make template compilation work when scaffolders are referenced as packages.
When scaffolder is used as a package reference, unless the app's code takes
some dependency on scaffolder code
(which is very unlikely for target app for which scaffolding is being done),
the scaffolding framework assemblies are not in app's closure of assemblies,
so we pass in all references in the closure of framework.

However this model still breaks for custom scaffolders because that may be using a different closure.