Changes to edit startup while creating a new context.
This also fixes the issue with ApplicationDbContext. (The change to pass the Startup type to web host build fixed that issue).

Though I have added ContextProcessingStatus and ContextProcessingResult with the view
that, if we could get Model metadata but could not edit Startup, we need
that information to Controllers and Views generator to throw an exception.

However, when we could not edit Startup, getting model metadata itself does not work and throws
exceptions but I kept the code just in case it will be needed in future.