Move SelectList code-generation into OnPageHandlerExecutionAsync
`SelectList` initializers are only being code-generated into the GET
handlers, but not the POST handlers.  Since `.ViewData` entries do not
persist across responses, the page will not render correctly when the
POST handler does not redirect (i.e. because validation failed).

This change moves the initialization into a reusable
`OnPageHandlerExecutionAsync`, which is invoked after model binding but
before the handler itself.  No need to dual-generate the initializers
into both handlers!